corrections for mobile nav
