+++ contact.html
@@ -10,7 +10,7 @@
       <link rel="preconnect" href="https://fonts.googleapis.com">
       <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
       <link href="https://fonts.googleapis.com/css2?family=Chewy&family=Indie+Flower&family=Judson:ital,wght@0,400;0,700;1,400&family=Open+Sans:ital,wght@0,300..800;1,300..800&family=Poppins:ital,wght@0,100;0,200;0,300;0,400;0,500;0,600;0,700;0,800;0,900;1,100;1,200;1,300;1,400;1,500;1,600;1,700;1,800;1,900&display=swap" rel="stylesheet">
-      <script defer src="./javascript/contact.js" type="module"></script>
+      <script defer src="./javascript/shared.js" type="module"></script>
       <title>Pawsitivity Partners</title>
   </head>
   <body class="contact--page">
@@ -36,31 +36,33 @@
           <button class="navigation--button"><a href="contact.html">CONTACT US</a></button>
         </div>
       </nav>
-      <h1 class="contact--main-header">CONTACT US</h1>
     </header>
     <main>
-      <div class="contact-form">
-        <form action="/submit_contact_form" method="post">
-          <div class ="form-row">
-            <div class="form-column">
-              <label for="name">Name</label>
-              <input type="text" id="name" name="name" required>
-              <label for="email">Email</label>
-              <input type="email" id="email" name="email" required>
-              <label for="phone">Phone</label>
-              <input type="tel" id="phone" name="phone" required>
+      <h1 class="contact--main-header">CONTACT US</h1>
+      <div class="contact-form-container">
+        <div class="contact-form">
+          <form action="/submit_contact_form" method="post">
+            <div class ="form-row">
+              <div class="form-column">
+                <label for="name">Name</label>
+                <input type="text" id="name" name="name" required>
+                <label for="email">Email</label>
+                <input type="email" id="email" name="email" required>
+                <label for="phone">Phone</label>
+                <input type="tel" id="phone" name="phone" required>
+              </div>
+            <div class = "form-column">
+              <label for="date">Date</label>
+              <input type="text" id="date" name="date">
+              <label for="message">Message</label>
+              <textarea id="message" name="message" required></textarea>
             </div>
-          <div class = "form-column">
-            <label for="date">Date</label>
-            <input type="text" id="date" name="date">
-            <label for="message">Message</label>
-            <textarea id="message" name="message" required></textarea>
           </div>
+          <button type="submit">Send</button>
+          </form>
         </div>
-        <button type="submit">Send</button>
-        </form>
       </div>
-      
+
       <div class="footer">
         <div class="left-text">
           <p class="bold">Want to reach out to us?</p>
@@ -74,7 +76,7 @@
             <a href="#" class ='socialmedia-icon'><img src="images/instagram.svg" alt="Instagram"></a>
           </div>
         </div>
-      </div>   
+      </div>
     </main>
   </body>
 </html>
