+++ stylesheets/contact.css
@@ -1,7 +1,7 @@
 body {
   background-color: #2F97C1;
-  padding: 20px;
-  overflow: hidden;
+  /* padding: 20px; */
+  /* overflow: hidden; */
 }
 
 .navigation--button {
@@ -14,32 +14,36 @@ body {
   border-radius: 16px;
   color: white;
 }
+
 .nav--buttons-container {
   display: flex;
   width: 20rem;
   justify-content: space-between;
 }
 
+.contact-form-container {
+  margin: 8vh 4vw 0;
+}
 
 .contact-form {
   background-color: white;
   display: flex;
   flex-direction: column;
-  width:80%; 
-  max-width: 500px; 
-  margin: 150px auto 20px; 
+  /* width: 80%; */
+  max-width: 500px;
+  margin: 0 auto;
   padding: 80px;
   border-radius: 16px;
-  box-shadow: 0px 0px 10px rgba(0, 0, 0, 0.1); 
+  box-shadow: 0px 0px 10px rgba(0, 0, 0, 0.1);
 }
 
-.form-row{
+.form-row {
   display: grid;
-  grid-template-columns: 1fr 1fr;
-  gap: 30px;
+  grid-template-columns: 1fr;
+  /* gap: 30px; */
 }
 
-.form-column{
+.form-column {
   display: flex;
   flex-direction: column;
   font-family: var(--body);
@@ -52,19 +56,21 @@ body {
   font-size: 3rem;
   color: var(--yellow);
   letter-spacing: 0.2rem;
-  animation: slideIn 1s ease forwards;
-  position: fixed;
-  top: 100px;
-  left: 50%;
-  transform: translate(-50%, -50%);
+  text-align: center;
+  margin-top: 12vh;
+  /* animation: slideIn 1s ease forwards; */
+  /* position: fixed; */
+  /* top: 100px; */
+  /* left: 50%; */
+  /* transform: translate(-50%, -50%); */
 }
 
 textarea {
   border: 2px solid black;
-  padding: 20px; 
-  resize: none ; 
-  width:100%;
-  height: 32px;
+  padding: 0;
+  resize: none ;
+  width: 100%;
+  height: 80px;
 }
 
 input[type="text"],input[type="email"], input[type="tel"], input [type="date"] {
@@ -96,19 +102,23 @@ button:hover {
 
 .footer {
   display: flex;
+  flex-direction: column;
   align-items: center;
   justify-content: space-between; /* Ensure items are spaced evenly */
   text-align: center; /* Center align text within each div */
+  gap: 2rem;
   padding: 10px;
   background-color: #2F97C1;
   font-family: var(--body);
-  color:white
+  color: white;
+  margin: 2rem auto;
 }
 
 .social-media-icons {
   display: flex; /* Ensure the icons are displayed in a row */
   /* align-items: center; */
   justify-content: center;
+  align-items: center;
 }
 
 .socialmedia-icon {
@@ -137,7 +147,19 @@ button:hover {
 }
 
 .socialmedia-icon img {
-    width: 30px;
-    height:auto;
-    margin-top: 5px;
+  width: 30px;
+  height:auto;
+  margin-top: 5px;
+}
+
+@media (min-width: 600px) {
+  .form-row {
+    grid-template-columns: 1fr 1fr;
+    gap: 30px;
+  }
+
+  .footer {
+    flex-direction: row;
+    align-items: start;
+  }
 }
